Version 1.0
Added a check to make sure there's white-space separating
message pattern from the body of the method.  Makes
errors somewhat easier to spot.

Added a fix to allow use of line continuation.  Script will no
longer fail when 'compatible' is set.

Added support for spell checking.  It's now performed only
in comments and strings.

Numbers are now handled separately as integers, floats and
scaled decimals.  The patterns now conform to the
draft ANSI Smalltalk standard which means that some
GNU Smalltalk specific constructs are no longer supported.
It's planned to reintroduce those in a different fashion
some time "in the future".  Which is soon, hopefully.

Temporarily removed error highlighting from array literals.  This is a
side effect of the change above.

Symbol handling was also changed for conformance.  In
particular, Symbol literals that aren't strings must
be valid selectors.  This isn't quite true in
GNU Smalltalk, but it's also something that I'll add back
later.